// apps/nest-js/src/shared/sheet/cell/types.ts
export type MergeParams = {
    cellEnd?: string;
    cellStart?: string;
    positions?: {
        endRow: number;
        startRow: number;
        endColumn: number;
        startColumn: number;
    }
}

export type CellParams = {
    cell: string | number;
    cellColumn?: string | number;
    type?: TCell;
    value: string | number | boolean;
    merge?: MergeParams;
    styles?: Partial<CellStyles>;
}

export type CellStyles = {
    fill: FillStyles;
    fillColor?: string;
    font: FontStyles;
    fontColor?: string;
    border?: BorderStyles;
    alignment: AlignmentStyle;
    borderStyle?: TBorderStyle;
}

type FillStyles = {
    type: 'pattern';
    pattern: | 'none' | 'solid' | 'darkVertical' | 'darkHorizontal' | 'darkGrid' | 'darkTrellis' | 'darkDown' | 'darkUp' | 'lightVertical' | 'lightHorizontal' | 'lightGrid' | 'lightTrellis' | 'lightDown' | 'lightUp' | 'darkGray' | 'mediumGray' | 'lightGray' | 'gray125' | 'gray0625';
    fgColor?: ColorStyle;
}

type FontStyles = {
    name?: string;
    size?: number;
    bold?: boolean;
    color?: ColorStyle;
}

type ColorStyle = {
    argb: string;
};

type TCell = 'title' | 'text';

type BorderStyles = {
    top: { style: TBorderStyle; };
    left: { style: TBorderStyle; };
    bottom: { style: TBorderStyle; };
    right: { style: TBorderStyle; };
}

type AlignmentStyle = {
    vertical?: 'top' | 'middle' | 'bottom' | 'distributed' | 'justify';
    wrapText?: boolean;
    horizontal?: 'left' | 'center' | 'right' | 'fill' | 'justify' | 'centerContinuous' | 'distributed';
}

type TBorderStyle = | 'thin' | 'dotted' | 'hair' | 'medium' | 'double' | 'thick' | 'dashed' | 'dashDot' | 'dashDotDot' | 'slantDashDot' | 'mediumDashed' | 'mediumDashDotDot' | 'mediumDashDot';


export type CellBorderStyle = | 'thin' | 'dotted' | 'hair' | 'medium' | 'double' | 'thick' | 'dashed' | 'dashDot' | 'dashDotDot' | 'slantDashDot' | 'mediumDashed' | 'mediumDashDotDot' | 'mediumDashDot';